functional nidaq_.py
The nidaq_ interface is now functional but I'm still not completely
happy with it. Currently the entire wave file is stored in memory and
transferred to buffer by nidaqmx in a separate thread. This seems okay
for normal sized files, but could be a problem with larger ones. Also,
we have no way of getting when the sound finished as of now.